# docs/swagger.yaml
definitions:
...
  dto.BannerItem:
    description: A promotional banner shown on the storefront.
    properties:
      banner_id:
        description: Unique identifier for the banner
        type: string
      created_at:
        description: Creation timestamp, RFC3339
        type: string
      end_date:
        description: Date the banner stops being visible, RFC3339
        type: string
      image_url:
        description: URL of the banner image
        type: string
      is_active:
        description: |-
          Manual on/off switch. Can only turn a banner OFF — it never forces one
          live outside its date window.
        type: boolean
      is_live:
        description: |-
          Derived, read-only: whether the banner is actually showing right now
          (is_active AND inside the date window). This is what the grid should display.
        type: boolean
      start_date:
        description: Date the banner becomes visible, RFC3339
        type: string
      title:
        description: Admin-facing name, so banners are distinguishable in the grid
        type: string
    type: object
...
  dto.Banners:
    description: Paginated list of banners.
    properties:
      banners:
        description: Page of banners
        items:
          $ref: '#/definitions/dto.BannerItem'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/query.Meta'
        description: Pagination metadata
    type: object
...
  dto.Cities:
    description: Paginated list of cities.
    properties:
      cities:
        description: Page of cities
        items:
          $ref: '#/definitions/dto.CityItem'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/query.Meta'
        description: Pagination metadata
    type: object
  dto.CityDTO:
    properties:
      city_id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  dto.CityItem:
    description: A serviceable city. The lat/lng + default_pincode are the location
    properties:
      city_id:
        description: Unique identifier for the city
        type: string
      default_pincode:
        description: Pincode used as the default QuickCommerce location anchor
        type: string
      enabled:
        description: Whether the city is serviceable
        type: boolean
      lat:
        description: Latitude of the city's QuickCommerce location anchor
        type: number
      lng:
        description: Longitude of the city's QuickCommerce location anchor
        type: number
      name:
        description: Display name of the city
        type: string
      slug:
        description: URL-safe name
        type: string
      state:
        description: State the city sits in
        type: string
    type: object
...
  dto.Response:
    description: Generic response envelope wrapping every API response.
    properties:
      code:
        description: Machine-readable error/status code, present on failures
        type: string
      data:
        description: Response payload, shape depends on the endpoint
      message:
        description: Human-readable message describing the result
        type: string
      status:
        description: Outcome of the request, e.g. "success" or "failed"
        type: string
    type: object
...
  query.Meta:
    properties:
      page:
        type: integer
      pages:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  v1.BanUserRequest:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  v1.CreateBannerRequest:
    properties:
      end_date:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      start_date:
        type: string
      title:
        type: string
    required:
    - image_url
    - title
    type: object
...
  v1.CreateCityRequest:
    properties:
      default_pincode:
        type: string
      enabled:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      state:
        type: string
    required:
    - default_pincode
    - lat
    - lng
    - name
    type: object
...
  v1.DeleteBannerRequest:
    properties:
      banner_id:
        type: string
    required:
    - banner_id
    type: object
...
  v1.DeleteCityRequest:
    properties:
      city_id:
        type: string
    required:
    - city_id
    type: object
...
  v1.UpdateBannerRequest:
    properties:
      banner_id:
        type: string
      end_date:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      start_date:
        type: string
      title:
        type: string
    required:
    - banner_id
    type: object
...
  v1.UpdateCityRequest:
    properties:
      city_id:
        type: string
      default_pincode:
        type: string
      enabled:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      state:
        type: string
    required:
    - city_id
    type: object
...
info:
  contact:
    name: API Support
  description: This is the API documentation for the Grocerics backend service.
  title: Grocerics API
  version: "1.1"
paths:
...
  /v1/banners:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete Banner Request
        in: body
        name: DeleteBannerRequest
        required: true
        schema:
          $ref: '#/definitions/v1.DeleteBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a banner
      tags:
      - banners
    get:
      description: Paginated list of all banners (admin sees inactive/expired too).
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Banners'
              type: object
      security:
      - BearerAuth: []
      summary: Get banners
      tags:
      - banners
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update Banner Request
        in: body
        name: banner
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BannerItem'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a banner
      tags:
      - banners
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Banner Request
        in: body
        name: banner
        required: true
        schema:
          $ref: '#/definitions/v1.CreateBannerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BannerItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a banner
      tags:
      - banners
  /v1/banners/{banner_id}:
    get:
      parameters:
      - description: Banner ID
        in: path
        name: banner_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BannerItem'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get banner by ID
      tags:
      - banners
...
  /v1/cities:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete City Request
        in: body
        name: DeleteCityRequest
        required: true
        schema:
          $ref: '#/definitions/v1.DeleteCityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a city
      tags:
      - cities
    get:
      description: Cities the app serves — used by the location picker.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CityDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List serviceable cities
      tags:
      - consumer
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update City Request
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateCityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CityItem'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a city
      tags:
      - cities
    post:
      consumes:
      - application/json
      parameters:
      - description: Create City Request
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/v1.CreateCityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CityItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a serviceable city
      tags:
      - cities
  /v1/cities/{id}:
    get:
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CityItem'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get a city by ID (admin)
      tags:
      - cities
  /v1/cities/all:
    get:
      description: Paginated list of serviceable cities, including disabled.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Cities'
              type: object
      security:
      - BearerAuth: []
      summary: List all cities (admin)
      tags:
      - cities
...
  /v1/users/ban:
    post:
      consumes:
      - application/json
      description: Sets the client's status to "banned". Irreversible.
      parameters:
      - description: Ban User Request
        in: body
        name: BanUserRequest
        required: true
        schema:
          $ref: '#/definitions/v1.BanUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User banned successfully
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Ban a client
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"